Added fog and fixed cubic voxel meshing colour bug

+++ src/libvolume/engine/realm.h
@@ -19,7 +19,7 @@ namespace LibVolume
 		class Realm
 		{
 			public:
-				long long time;
+				long time;
 
 				Camera camera;
